+++ src/services/PagesDomains.js
@@ -7,7 +7,7 @@ class PagesDomains extends BaseService {
   all(projectId) {
     const url = projectId ? `projects/${encodeURIComponent(projectId)}/` : '';
 
-    return RequestHelper.get(this, `${url}page/domains`);
+    return RequestHelper.get(this, `${url}pages/domains`);
   }
 
   @api('<projectId>', '<domain>', { options: true, method: 'POST' })